Data.Xls

Description

Parse Microsoft excel spreadsheet xls file (format BIFF/Excel 97-2004).

Documentation

decodeXls :: MonadResource m => FilePath -> Producer m [String]

Parse a Microsoft excel xls workbook file into a Conduit yielding rows in a worksheet. Each row represented by a list of Strings, each String representing an individual cell.

Currently there is no separation of worksheets, all worksheets in a workbook get concatenated.

Throws XlsException

data XlsException

Constructors

XlsFileNotFound String

XlsParseError String